optimization refactoring 2
